==> Notebooks/2_Supermarket_Extractor.ipynb <==
# Índice

**Extractores de URLs**

    1️⃣ Configuración y creación de lista de URLs  
    2️⃣ Extractor lácteos
    3️⃣ Extractor frescos   
    4️⃣ Extractor alimentación  
    5️⃣ Extractor bebidas   
    6️⃣ Extractor congelados 
    

**Extractor de contenido**

**Creación de fichero .csv**

**Anexo: Listado de marcas**


## Extractores de URLs

### 1. Configuración y creación de lista de URLs

El planteamiento de la extracción de las Url de productos después de estudiar la Web del supermercado es seleccionar sólo aquellas categorías que pueden aportar información relevante al objetivo del análisis (descartando categorías como productos de limpieza, higiene..). Además se extraerán las URL de los productos de las categorías por separado, dado que hay diferencias en algunos elementos de las mismas (como el número inicial de productos a mostrar por página).

El funcionamiento de los extractores de las categorías consta de dos pasos; tras abrir la web se ejecuta un primer bucle en el que se localiza cada uno de los botones que despliegan el listado de productos hasta tenerlo completo, una vez finalizada esa acción se ejecuta un segundo bucle que localiza los link de cada uno de los productos. En este ejemplo el número de elementos que se incluyen en el segundo bucle se basa en la información de la página en el día concreto de esta extracción, para estandarizarlo se puede fijar una cifra superior para tener un margen (ya que el número de productos por categoría varía de forma diaria y he visto que las etiquetas que muestran cuántos productos hay en cada una de ellas en de la web no son 100% exactas)

El resultado de la acción anterior es conseguir las URL de todos los productos de cada categoría para poder acceder a ellos y extraer la información necesaria en el siguiente punto sin tener que hacer cilck en todos. 

import pandas as pd
import numpy as np

import random
from random import randint
import string
import requests
from tqdm import tqdm

from time import sleep
from selenium import webdriver
from bs4 import BeautifulSoup
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.chrome.options import Options

import warnings
warnings.filterwarnings('ignore')

opciones= Options()
opciones.add_experimental_option('excludeSwitches', ['enable-automation'])
#para ocultarme como robot
opciones.add_experimental_option('useAutomationExtension', False)
opciones.add_argument('--start-maximized') #empezar maximizado
opciones.add_argument('user.data-dir=selenium') #guarda las cookies
opciones.add_argument('--incognito')#incognito window

url_productos=[]

### 2. Extractor lácteos

driver = webdriver.Chrome("chromedriver.exe")

categoria= 'lacteos'
num_res= list(range(15, 5000, 12))  


driver.get(f"https://www.ahorramas.com/{categoria}/")

try:
    driver.find_element("xpath", "//*[@id='onetrust-accept-btn-handler']").click()
    sleep(7)
except:
    pass

for num in tqdm(num_res):
    try:
        driver.find_element("xpath", f'//*[@id="product-search-results"]/div[2]/div[4]/div[2]/div[{num}]/div/div/button').click()
        driver.implicitly_wait(60)
        sleep(randint(2,5))
    except:
        print(f"No more Show more button: {num}")
        break
        
for i in tqdm(range(1,439)):
    try:
        url_productos.append(driver.find_element("xpath", f"//*[@id='product-search-results']/div[2]/div[4]/div[2]/div[{i}]/div/div/div[1]/a").get_attribute('href'))
        
    except:
        print(f"no such product: {i}")
        pass

### 3. Extractor frescos

driver = webdriver.Chrome("chromedriver.exe")

categoria= 'frescos'
num_res= list(range(13, 5000, 12)) 


driver.get(f"https://www.ahorramas.com/{categoria}/")

try:
    driver.find_element("xpath", '//*[@id="onetrust-accept-btn-handler"]').click()
    sleep(7)
except:
    pass

for num in tqdm(num_res):
    try:
        driver.find_element("xpath", f'//*[@id="product-search-results"]/div[2]/div[4]/div[2]/div[{num}]/div/div/button').click()
        driver.implicitly_wait(60)
        sleep(randint(2,5))
    except:
        print(f"No more Show more button: {num}")
        break
        
for i in tqdm(range(1,1045)):
    try:
        url_productos.append(driver.find_element("xpath", f"//*[@id='product-search-results']/div[2]/div[4]/div[2]/div[{i}]/div/div/div[1]/a").get_attribute('href'))
        
    except:
        print(f"no such product: {i}")
        pass

### 4. Extractor alimentación

driver = webdriver.Chrome("chromedriver.exe")

categoria= 'alimentacion'
num_res= list(range(13, 5000, 12)) 


driver.get(f"https://www.ahorramas.com/{categoria}/")

try:
    driver.find_element("xpath", '//*[@id="onetrust-accept-btn-handler"]').click()
    sleep(7)
except:
    pass

for num in tqdm(num_res):
    try:
        driver.find_element("xpath", f'//*[@id="product-search-results"]/div[2]/div[4]/div[3]/div[{num}]/div/div/button').click()
        driver.implicitly_wait(60)
        sleep(randint(2,5))
    except:
        print(f"No more Show more button: {num}")
        break
  

for i in tqdm(range(1,1945)):
    try:
        url_productos.append(driver.find_element("xpath", f"//*[@id='product-search-results']/div[2]/div[4]/div[3]/div[{i}]/div/div/div[1]/a").get_attribute('href'))
        
    except:
        print(f"no such product: {i}")
        pass

### 5. Extractor bebidas

driver = webdriver.Chrome("chromedriver.exe")

categoria= 'bebidas'
num_res= list(range(15, 5000, 12)) 


driver.get(f"https://www.ahorramas.com/{categoria}/")

try:
    driver.find_element("xpath", '//*[@id="onetrust-accept-btn-handler"]').click()
    sleep(7)
except:
    pass

for num in tqdm(num_res):
    try:
        driver.find_element("xpath", f'//*[@id="product-search-results"]/div[2]/div[4]/div[3]/div[{num}]/div/div/button').click()
        driver.implicitly_wait(60)
        sleep(randint(2,5))
    except:
        print(f"No more Show more button: {num}")
        break
        
for i in tqdm(range(1,620)):
    try:
        url_productos.append(driver.find_element("xpath", f"//*[@id='product-search-results']/div[2]/div[4]/div[3]/div[{i}]/div/div/div[1]/a").get_attribute('href'))
        
    except:
        print(f"no such product: {i}")
        pass

### 6. Extractor congelados

driver = webdriver.Chrome("chromedriver.exe")

categoria= 'congelados'
num_res= list(range(14, 5000, 12)) 


driver.get(f"https://www.ahorramas.com/{categoria}/")

try:
    driver.find_element("xpath", '//*[@id="onetrust-accept-btn-handler"]').click()
    sleep(7)
except:
    pass

for num in tqdm(num_res):
    try:
        driver.find_element("xpath", f'//*[@id="product-search-results"]/div[2]/div[4]/div[2]/div[{num}]/div/div/button').click()
        driver.implicitly_wait(60)
        sleep(randint(2,5))
    except:
        print(f"No more Show more button: {num}")
        break
        
for i in tqdm(range(1,325)):
    try:
        url_productos.append(driver.find_element("xpath", f"//*[@id='product-search-results']/div[2]/div[4]/div[2]/div[{i}]/div/div/div[1]/a").get_attribute('href'))
        
    except:
        print(f"no such product: {i}")
        pass

df_url= pd.DataFrame(url_productos)

df_url.columns = ['URL']

df_url.to_csv('0_Proyecto_Recetas/Data/0_Others_data/1_LISTADO_URL_PRODUCTOS_AHORRAMAS.csv', header=True, index=False)

## Extractor de contenido

En base a la lista de Url generada en el punto anterior vamos a extraer en el siguiente código la información de cada uno de los productos. El procedimiento tarda en torno a 9 horas por norma general.

pd_lista=pd.read_csv('0_Proyecto_Recetas/Data/0_Others_data/1_LISTADO_URL_PRODUCTOS_AHORRAMAS.csv')

url_productos=[]

for url in pd_lista['URL']:
    url_productos.append(url)

sopa_ini=[]

for url in tqdm(url_productos):
    sopa_ini.append(requests.get(url))
    sleep(randint(5,9))

sopa=[]

for s in tqdm(sopa_ini):
    sopa.append(BeautifulSoup(s.content, 'html.parser'))

## Creación de fichero .csv

Una vez tenemos la información de los productos, localizaremos y almacenaremos en un diccionario lo siguiente:

    - producto_desc: descripción o nombre del producto   
    - precio_ud: precio del producto
    - precio_medida: precio por unidad de medida (kg, l...)   
    - peso_unit: peso aproximado de la unidad 
    - descuento: creado manualmente como resultado de una condición (valores: sí o no)   
    - precio_sin_desc: en caso de no tener descuento valor precio unitario, de lo contrario el precio sin descuento 
    - categorias: las diferentes categorías con las que el supermercado clasifica el producto

Generamos también como paso previo una lista de valores excluidos, que contiene los mensajes que muestra la web cuando una url no está disponible.

dicc_sup_ahorramas= {"producto_desc"     : [],
                      "precio_ud"        : [],
                      "precio_medida"    : [],
                      "peso_unit"        : [],
                      "descuento"        : [],
                      "precio_sin_desc"  : [],
                      "categorias"       : []                                
                     }

lista_excluidos=['URL Not Found','We cannot find the page you are looking for!']

for producto in tqdm(sopa):
    
    try:
        if ((producto.find_all(("h1", {"class": "product-name d-sm-none"}))[0].getText())) in lista_excluidos:
            pass
        
        else:
    
            dicc_sup_ahorramas["producto_desc"].append(producto.find_all(("h1", {"class": "product-name d-sm-none"}))[0].getText())

            try:
                dicc_sup_ahorramas["precio_ud"].append(producto.find_all("span", {"class": "value"})[0].getText())
                
            except:
                dicc_sup_ahorramas["precio_ud"].append(np.nan)
                
            try:
                dicc_sup_ahorramas["peso_unit"].append(producto.find_all("p", {"class": "font-small-12 mb-1"})[0].getText())
                
            except:
                dicc_sup_ahorramas["peso_unit"].append(np.nan)
                
            try:
                dicc_sup_ahorramas["categorias"].append(producto.find_all(("ol", {"class": "breadcrumb"}))[0].getText())
                
            except:
                
                dicc_sup_ahorramas["categorias"].append(np.nan)


            try:
                dicc_sup_ahorramas["precio_medida"].append(producto.find_all("span", {"class": "unit-price-per-unit grey"})[0].getText())
                dicc_sup_ahorramas["descuento"].append("No")
                dicc_sup_ahorramas["precio_sin_desc"].append("precio unitario")
                
                continue
                
            except:
                
                try:
                    dicc_sup_ahorramas["precio_medida"].append(producto.find_all("span", {"class": "unit-price-per-unit red"})[0].getText())
                    dicc_sup_ahorramas["descuento"].append("Sí")
                    dicc_sup_ahorramas["precio_sin_desc"].append(producto.find_all("span", {"class": "unit-price-old"})[0].getText()+' ('+
                                                             producto.find_all("span", {"class": "marker"})[0].getText()+')')
                   
                    continue
                    
                except:
                    
                    dicc_sup_ahorramas["precio_medida"].append(np.nan)
                    dicc_sup_ahorramas["descuento"].append(np.nan)
                    dicc_sup_ahorramas["precio_sin_desc"].append(np.nan)
                    
                    continue

                continue

    except:
        pass

#Comprobamos la longitud de cada una de las key de nuestro diccionario

print(len(dicc_sup_ahorramas["precio_medida"]))
print(len(dicc_sup_ahorramas["descuento"]))
print(len(dicc_sup_ahorramas["precio_sin_desc"]))
print(len(dicc_sup_ahorramas["producto_desc"]))
print(len(dicc_sup_ahorramas["precio_ud"]))
print(len(dicc_sup_ahorramas["peso_unit"]))
print(len(dicc_sup_ahorramas["categorias"]))

df_catalogo= pd.DataFrame(dicc_sup_ahorramas)

df_catalogo.to_csv('0_Proyecto_Recetas/Data/2_Extracted_data/3_CATALOGO_AHORRAMAS_DATA.csv', header=True, index=False)

## Anexo: Listado de marcas

A lo largo de la limpieza de los datos ha surgido la necesidad de eliminar las marcas de los proveedores de los nombres de los productos, por lo que se ha incluido este anexo que obtiene las marcas por cada una de las categorías estudiadas.

lista_marcas=[]

driver = webdriver.Chrome("chromedriver.exe")

categorias= ['lacteos','congelados','bebidas','alimentacion','frescos']

for categoria in categorias:
    driver.get(f"https://www.ahorramas.com/{categoria}/")
    
    
    try:
        driver.find_element("xpath", '//*[@id="onetrust-accept-btn-handler"]').click()
        sleep(randint(2,5))
    except:
        pass
    
    driver.find_element("xpath", '//*[@id="refinement-marca"]/div[1]/span').click()
    sleep(randint(13,15))
    
    for i in tqdm(range(0,500)):
        try:
            lista_marcas.append(driver.find_element("xpath", f'//*[@id="refinement-marca"]/ul[1]/li[{i}]/button/span[1]').text)
        except:
            
            try:
                lista_marcas.append(driver.find_element("xpath", f'//*[@id="refinement-marca"]/ul[2]/li[{i}]/button/span[1]').text)
                continue
            except:
                pass

    sleep(randint(4,9))

df_marcas= pd.DataFrame(lista_marcas)

df_marcas.columns = ['MARCA']

df_marcas.to_csv('../Data/0_Others_data/3_LISTADO_MARCAS_AHORRAMAS.csv', header=True, index=False)
