==> Notebooks/3_Recipes_Cleaning.ipynb <==
# Índice

**Limpieza inicial dataframe**

    1️⃣ Borrado de columnas
    2️⃣ Traducción columnas de texto

**Tratamiento de columnas**

    1️⃣ Añadir columna de instrucciones
    2️⃣ Manejo columnas de tiempo

**Creación de fichero .csv**

import pandas as pd
import numpy as np
import re
import string
from deep_translator import GoogleTranslator

import sys 
sys.path.append("../SRC")
import support as sp 

df=pd.read_csv("../Data/2_Extracted_data/2_RECIPES_DATA.csv")
df_instrucciones=pd.read_csv("../Data/2_Extracted_data/4_INSTRUCTIONS_DATA.csv")
pd.options.display.max_columns = None

## Limpieza inicial dataframe

**1. Borrado de columnas**

sp.Limpieza_columnas_general(df,"AuthorId").borrado_columna()
sp.Limpieza_columnas_general(df,"AuthorName").borrado_columna()
sp.Limpieza_columnas_general(df,"CookTime").borrado_columna()
sp.Limpieza_columnas_general(df,"DatePublished").borrado_columna()
sp.Limpieza_columnas_general(df,"Images").borrado_columna()
sp.Limpieza_columnas_general(df,"Keywords").borrado_columna()
sp.Limpieza_columnas_general(df,"RecipeIngredientQuantities").borrado_columna()
sp.Limpieza_columnas_general(df,"RecipeIngredientParts").borrado_columna()
sp.Limpieza_columnas_general(df,"RecipeServings").borrado_columna()
sp.Limpieza_columnas_general(df,"RecipeYield").borrado_columna()
sp.Limpieza_columnas_general(df,"Description").borrado_columna()
sp.Limpieza_columnas_general(df,"RecipeInstructions").borrado_columna()

**2. Traducción columnas de texto**

sp.Limpieza_columnas_general(df,"RecipeCategory").series_a_string()
sp.Limpieza_columnas_general(df,"Name").series_a_string()

sp.Limpieza_columnas_general(df,"RecipeCategory").funcion_traduccion_en_es("Categoria_receta")
sp.Limpieza_columnas_general(df,"Name").funcion_traduccion_en_es("Nombre")

sp.Limpieza_columnas_general(df,"RecipeCategory").borrado_columna()
sp.Limpieza_columnas_general(df,"Name").borrado_columna()

## Tratamiento de columnas

**1. Añadir columna de instrucciones**

sp.Limpieza_columnas_general(df_instrucciones,"url").busqueda_patrones_regex("RecipeId","\-\d{4,20}")
df_instrucciones["RecipeId"] = df_instrucciones["RecipeId"].explode()
sp.Limpieza_columnas_general(df_instrucciones,"RecipeId").reemplazo_string("-",'')
sp.Limpieza_columnas_general(df_instrucciones,"url").borrado_columna()

df_instrucciones=df_instrucciones.groupby('RecipeId')['intrucciones'].apply(list)

df_instrucciones = pd.DataFrame(df_instrucciones)

df_instrucciones["intrucciones"] = [','.join(map(str, i)) for i in df_instrucciones["intrucciones"]]

df_instrucciones.reset_index(drop=False, inplace=True)

df_instrucciones["RecipeId"]=df_instrucciones["RecipeId"].astype("int")

df = pd.merge(left=df, right=df_instrucciones, how='left',
                        left_on='RecipeId', right_on='RecipeId')

sp.Limpieza_columnas_general(df,"intrucciones").funcion_traduccion_en_es("intrucciones")

**4. Manejo columnas de tiempo**

sp.Limpieza_columnas_general(df,"PrepTime").reemplazo_string("PT",'')
sp.Limpieza_columnas_general(df,"TotalTime").reemplazo_string("PT",'')

Transformación PrepTime a minutos:

df=df.join(df["PrepTime"].str.split("H", expand=True)).rename(columns={0:'PrepTime_1', 1:'PrepTime_2'})

sp.Limpieza_columnas_general(df,"PrepTime_2").reemplazo_valores_nulos("MINUTOS")

df["minutos"] = np.where((df["PrepTime_2"]=="MINUTOS"), df["PrepTime_1"], df["PrepTime_2"])
df["horas"] = np.where((df["PrepTime_2"]=="MINUTOS"), "0", df["PrepTime_1"])

df["minutos"] = np.where((df["minutos"]==""), "0", df["minutos"])

sp.Limpieza_columnas_general(df,"PrepTime_1").borrado_columna()
sp.Limpieza_columnas_general(df,"PrepTime_2").borrado_columna()

sp.Limpieza_columnas_general(df,"minutos").reemplazo_string("M",'')
sp.Limpieza_columnas_general(df,"minutos").reemplazo_string("0S",'0')

sp.Limpieza_columnas_general(df,"minutos").conversion_formato_columna("int")
sp.Limpieza_columnas_general(df,"horas").conversion_formato_columna("int")

df["Prep_time"]= df["minutos"]+(df["horas"]*60)

sp.Limpieza_columnas_general(df,"minutos").borrado_columna()
sp.Limpieza_columnas_general(df,"horas").borrado_columna()
sp.Limpieza_columnas_general(df,"PrepTime").borrado_columna()

Transformación TotalTime a minutos:

df=df.join(df["TotalTime"].str.split("H", expand=True)).rename(columns={0:'TotalTime_1', 1:'TotalTime_2'})

sp.Limpieza_columnas_general(df,"TotalTime_2").reemplazo_valores_nulos("MINUTOS")

df["minutos"] = np.where((df["TotalTime_2"]=="MINUTOS"), df["TotalTime_1"], df["TotalTime_2"])
df["horas"] = np.where((df["TotalTime_2"]=="MINUTOS"), "0", df["TotalTime_1"])

df["minutos"] = np.where((df["minutos"]==""), "0", df["minutos"])

sp.Limpieza_columnas_general(df,"TotalTime_1").borrado_columna()
sp.Limpieza_columnas_general(df,"TotalTime_2").borrado_columna()

sp.Limpieza_columnas_general(df,"minutos").reemplazo_string("M",'')

sp.Limpieza_columnas_general(df,"minutos").conversion_formato_columna("int")
sp.Limpieza_columnas_general(df,"horas").conversion_formato_columna("int")

df["Total_time"]= df["minutos"]+(df["horas"]*60)

sp.Limpieza_columnas_general(df,"minutos").borrado_columna()
sp.Limpieza_columnas_general(df,"horas").borrado_columna()
sp.Limpieza_columnas_general(df,"TotalTime").borrado_columna()

## Creación de fichero .csv

df.info()

df.to_csv('../Data/3_Clean_data/2_RECIPES_CLEAN.csv', header=True, index=False)
