==> Notebooks/6_MONGO_Database.ipynb <==
import pandas as pd
import pymongo

df_recetas=pd.read_csv("../Data/3_Clean_data/2_RECIPES_CLEAN.csv")
df_ingredientes=pd.read_csv("../Data/3_Clean_data/1_INGREDIENTS_CLEAN.csv")
df_supermercado=pd.read_csv("../Data/3_Clean_data/3_SUPERMARKET_CLEAN.csv")

# Conect with Mongo

client = pymongo.MongoClient()
client

# Check our databases

client.list_database_names()

# Create a new database on MongoDB


mydb = client["bbdd_etl_recipes"]

# Create a brand new collection in our database

collection_1 = mydb.recipes
collection_2 = mydb.ingredients
collection_3 = mydb.supermarket

# Insert data from our dataframe

collection_1.insert_many(df_recetas.to_dict("records"))
collection_2.insert_many(df_ingredientes.to_dict("records"))
collection_3.insert_many(df_supermercado.to_dict("records"))

# Pass the query parameter without any specification, so it will return all the elements of the database

collection.find()
